feat: US-07 - Extract snapshot module from log_payload.rs
